+++ app/src/tabs/index.ts
@@ -1,2 +1,4 @@
-export { default as Tabs, TabsList, TabsTrigger, TabsContent } from './Tabs';
-export type { TabsProps, TabsListProps, TabsTriggerProps, TabsContentProps } from './Tabs';
+export { Tabs, type TabsProps } from './Tabs';
+export { TabsContent, type TabsContentProps } from './TabsContent';
+export { TabsList, type TabsListProps } from './TabsList';
+export { TabsTrigger, type TabsTriggerProps } from './TabsTrigger';
